+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Version 194:
 
 * http::async_read returns the right byte count on error
 * Add net namespace alias
+* Don't use-after-free in test
+* Tidy up ssl_stream (experimental)
 
 --------------------------------------------------------------------------------